+++ terminal/connimpls/vt400/SerialConnVT420.cpp
@@ -54,7 +54,7 @@ void SerialConnVT420::ProcessDECDC(const std::string& p)
     if (bot < 0)
         bot = (int)mLines.size() - 1;
     for (int k = top; k <= bot; ++k) {
-        if (pn <= mLines[k].size()) {
+        if (pn <= (int)mLines[k].size()) {
             mLines[k].erase(mLines[k].begin() + pn - 1);
             mLines[k].push_back(mBlankChar);
         }
@@ -69,7 +69,7 @@ void SerialConnVT420::ProcessDECIC(const std::string& p)
     if (bot < 0)
         bot = (int)mLines.size() - 1;
     for (int k = top; k <= bot; ++k) {
-        if (pn <= mLines[k].size()) {
+        if (pn <= (int)mLines[k].size()) {
             mLines[k].insert(mLines[k].begin() + pn - 1, mBlankChar);
         }
     }
